Add read_param to config
